requirements.txt
tweepy
requests
beautifulsoup4
transformers
apscheduler
python-dotenv
Flask
